Update for compatibility with Babel 7.x
Babel 7.x made several changes in both their APIs and packaging
structure. This change updates babel-plugin-dedent to be compatible with
these changes.

The major API change which impacted the dedent plugin was a change to
the visitor function. In 6.x, this function was changed to no longer
need to return a value, but in 7.x, returning a value is treated as an
error. The simple fix for this is to remove the return statements.
This is backwards compatible with users who would still like to use the
plugin with Babel 6.x.

For the packaging changes, they've mostly just changed the names to
enforce a consistent standard. The old ES20.. presets are gone in favor
of `preset-env`, which attempts to be smart about selecting a set of
plugins that are useful given a target environment. I've updated the
devDependencies to pull in Babel 7.x, and changed the test suite to run
both against the env preset as well as the specific list of transform
plugins to target ES2015.

Test suite has been run against both Babel 6.x (using the old
package.json and test.js) and 7.x.